third_party/WebKit/LayoutTests/battery-status/test-battery-status-basic.html

<html>
    <body>
        <p>This test prints the default battery status values.</p>
        <p id="p1"></p>
        <p id="p2"></p>
        <p id="p3"></p>
        <p id="p4"></p>
    </body>
    <script>
    if (window.testRunner)
        testRunner.dumpAsText();
    
    var battery = navigator.battery;

    document.getElementById("p1").innerHTML = "Charging Status: " + (battery.charging ? "Charging" : "Discharging");
    document.getElementById("p2").innerHTML = "Charging Time: " + battery.chargingTime;
    document.getElementById("p3").innerHTML = "Discharging Time: " + battery.dischargingTime;
    document.getElementById("p4").innerHTML = "Percentage: " + battery.level*100;
    </script>
</html>
