third_party/WebKit/LayoutTests/battery-status/test-battery-status-basic-expected.txt

This test prints the default battery status values.

Charging Status: Charging

Charging Time: Infinity

Discharging Time: Infinity

Percentage: 100
